refactor: 🔥 Remove get_all_networks as dupe of get_all_subnets
The get_all_networks and get_all_subnets functions were doing the same query with the exception of one taking a prefix as a kwarg. I've combined these two functions and updated the references so only get_all_subnets is used and it supports defining a prefix.

Removed get_all_networks function.