fix: 🐛 Fix missing links for updates
It was found that during an update run that linked objects could be missed if the Custom Fields have the data filled in even if the linked object wasn't found. This will reset those fields so that the diff lines up correctly and will force an update to link those missing objects.