feat: Add Extensibility Attributes (#85)
* docs: :memo: Remove strikethrough on VLAN&Group
* chore: 🚚 Move images to doc static folder
* chore: 🚚 Move utility files into appropriate folder

Moved utility files into folder named the same. Also consolidated utility methods into single file. Updated all imports and references.

* build: Bump project to 0.5.2
* feat: ✨ Add Extensibility Attributes to each base model attributes
* refactor: 🔥 Remove get_all_networks as dupe of get_all_subnets

The get_all_networks and get_all_subnets functions were doing the same query with the exception of one taking a prefix as a kwarg. I've combined these two functions and updated the references so only get_all_subnets is used and it supports defining a prefix.

Removed get_all_networks function.

* style: 🚨 Remove whitespace for black
* fix: 🐛 Set prefix to None so not required get_all_subnets

Forgot to set prefix to None so it isn't a required kwarg. This should fix the tests that were failing.

* feat: ✨ Added method to standardize Extensibility Attributes dict and test
* feat: ✨ Update adapters to load Extensibility Attributes/Custom Fields for diff
* feat: ✨ Update client functions to get Extensibiilty Attributes
* feat: ✨ Create function to process Extensibility Attributes for each model

Wrote a function for each model called process_ext_attrs that reviews all EAs passed and attempts to match them with other objects in Nautobot like VRF, Site, Previx/VLANRole, and Tenant. Then it will also create CustomFields on each object for each EA and record values there.

* docs: Fix image links to docs folder
* refactor: 🚚 Rename utilities to diffsync to clarify use

As these functions are for both Infoblox and Nautobot the name diffsync makes more sense to align with use. As we expand these moving to individual files for each would be warranted. This makes the naming with the util folder more sensible.

Rename utilities to diffsync so paths are changed and location of functions needs to be updated.

* fix: 🐛 Ensure that update/delete functions use object pk

All models have a pk attribute to store the ID of an object from Nautobot. We should always use this for the update and delete functions on these models to ensure we are always working with the correct object.

* revert: ⏪️ Revert change in #78

I was wrong, the Tag objects are accepted. I'm not sure why previously I had some error thrown for the way it was done. I've tested and confirmed that the old method still works with 1.3 so it must have been a fluke.

* fix: Make key a slug ot match CustomFields
* build: Bump to 0.6.0

Should have been a minor bump, not patch as new features added.

* docs: 📝 Add documentation around Extensibility Attributes
* refactor: Change Site to OrmSite
* refactor: ♻️ Extracted process_ext_attrs function into helper function

To reduce repeated code I've moved the process_ext_attrs function outside the models and updated it to work with all of them. Logging had to be simplified but should still provide enough information to fix what's wrong if needed.

* fix: 🐛 Fix missing links for updates

It was found that during an update run that linked objects could be missed if the Custom Fields have the data filled in even if the linked object wasn't found. This will reset those fields so that the diff lines up correctly and will force an update to link those missing objects.

* fix: 🐛 Ensure that Role for IPAddress uses RoleChoices

To ensure that we don't accidentally have a Prefix or VLAN try to use a Role that IPAddress has I've added a check to ensure that the object passed is an IPAddress.

Co-authored-by: Justin Drew <jdrew82@users.noreply.github.com>