release: Enable mapping of VLANs to Networks
* feat: ✨ Add VLANs attribute to Network to allow pairing

To support VLANs assigned to a subnet I've added the VLANs attribute. I've also created a function to standardize the VLAN mapping.

* fix: Ensure CustomField value is a string

* build: Bump project to 0.6.1

* refactor: ♻️ Refactored the build_vlan_map function to include group

To make it easier to find the appropriate VLAN we need the Group or View when we're locating it.

* feat: ✨ Add Subnet to VLAN mapping

Will attempt to assign a VLAN to a Prefix if found. If more than one is found, it will create a Relationship with the additional VLANs.

* fix: 🐛 Move VLAN and VLANGroup in front of IPAM objects in top_level

In order to ensure that the VLANs are found for mapping to a Subnet they need to be created first. Making them the first objects in top level will ensure they're processed first before the Subnets and IPAddresses.

* build: Bump project to 0.7.0

* refactor: ♻️ Remove redundant syntax

Co-authored-by: Justin Drew <jdrew82@users.noreply.github.com>