fix: Fix Extensibility Attributes Churn & Fix Prefix VLAN Updates (#111)
* docs: 📝 Add docstring for conn variable

* fix: 🐛 Ensure Extensible Attributes match in diff. Fixes #108.

I've updated both adapters to ensure that the default CustomFields/Extensibility Attributes are loaded for all models. This ensures that he diff now lines up and reduces unnecessary changes from being attempted.

* style: 🚨 Fix variable name for pylint

* fix: 🐛 Correct Prefix VLAN updates. Fixes #110.

It was found that VLANs attached to Prefixes weren't being updated correctly. This corrects it and ensures that VLANs are attached and removed when necessary.

* test: 🚨 Fix pylint complaint

* style: 🚚 Rename variable to be more accurate to use

Co-authored-by: Justin Drew <jdrew82@users.noreply.github.com>