fix: IP Address Bulk Import (#118)
* fix: 🐛 Handle loading ip addresses from large datasets over 1000. Fixes #115.

It was found that on datasets where you're attempting to get more than a thousand devices the API will return an error response and no content. This refactors it to hopefully handle these cases by trying to anticipate large number of addresses being returned and splitting the query up accordingly.

* fix: ♻️ Refactor method to ensure there is something to log and return.

* fix: ♻️ Refactor ipaddr list creation to check list has contents.

This should ensure that empty list isn't added to the list.

* fix: 🐛 Ensure DNS name for IP Address are valid.

If invalid FQDNs are put in the dns_name attribute for an IP Address then it causes a ValidationError during IP Address creation. This corrects that while also updating the load to allow for IP Addresses without DNS assigned to be imported. I've also included a test for the method I created to make sure it works as expected.

* fix: 🐛 Catch non-existant VLAN throwing DoesNotExist

* test: 🚨 Fix pylint complaint

* perf: 🔇 Change IP Address logging to debug

* fix: ♻️ Final refactor of how IP addresses are bulk requested from API.

This should ensure that there is an optimal bulk request from the API without triggering the 1000 max result limit. I've also included tests of various response and network sizes to fully vet the operation of the function.

* test: 🚨 Fix docstring for pydocstyle

* style: Add logging during load of both adapters to give how many objects are loaded when done.

* fix: 🐛 Add try/except block around IP address create/update.

To catch Validation Errors I've added try/except blocks. This should hopefully allow for us to continue importing regardless of why an IP Address failing Nautobot's checks.

* test: 🚨 Fix ocmplaints from black

* fix: Check that objects were loaded before logging

* Update Nautobot adapter logging to validate key

* docs: 📝 Add docstrings and type hints to nested functions.

* perf: ⚡️ Refactor to use maps and bulk create method for all objects.

This should help address importing large data sets like hundreds of thousands of IP addresses.

* fix: 🐛 Update loaded VLANs to only be ones with associated Group.

To match how Infoblox requires VLANs to be associated to a VLANView we should only load those that have a Group associated. This should avoid potentially grabbing individual VLANs from other SoRs or manually input.

* refactor: ♻️ Change to filter to only those VLANs with groups.

No need to go through all VLANs to load them! We only need the ones with groups assigned. This should make it much quicker when there's a large amount of VLANs already in place.

* fix: 🐛 Use VLANGroup map and not ORM call when creating VLAN.

Looks like I missed updating a call to using the maps.

* refactor: ♻️ Remove requirement for DNS name for IP Address to be loaded.

* fix: 🐛 Remove id attribute from found_vlan since we're using map now.

* fix: 🔥 Remove unnecessary checks for VLANs and cleans

* docs: 📝 Add comment around why filter VLANs associated to Groups.

* docs: 📝 Update comment to explain max result limit for IP address requests.

* feat: ✨ Add plugin setting for default Status on Prefix and IP Addresses.

Co-authored-by: Justin Drew <jdrew82@users.noreply.github.com>